requirements.txt
# ThreatPilot - External Dependencies
# Python 3.11+

# UI framework
PySide6>=6.5

# Data validation
pydantic>=2.0

# Image handling
Pillow>=10.0

# OCR (optional)
pytesseract>=0.3.10

# AI communication
httpx>=0.25

# Storage
# sqlite3 is part of Python stdlib

# Export
pandas>=2.0
openpyxl>=3.1
